added polymorphic inheritance, reorganized react components, about to add like functionality to films and reviews

+++ app/controllers/api/reviews_controller.rb
@@ -1,23 +1,18 @@
-
-
 class Api::ReviewsController < ApplicationController
-  # def create
-  #   @review = Review.new(review_params)
-  #   if @review.save
-  #     render :show
-  #   else
-  #     render json: @review.errors.full_messages, status: 422
-  #   end
-  # end
+  before_action :require_login, only: [:update, :destroy]
 
   def update
     r = review_params.transform_keys(&:underscore).transform_values(&:presence)
-    @review = Review.find_by(user_id: r[:user_id],film_id: r[:film_id])
-    @review ? @review.update(r) : @review = Review.new(r)
-    if @review.save
-      render :show
+    if Integer(r[:user_id]) == current_user.id
+      @review = Review.find_by(user_id: r[:user_id],film_id: r[:film_id])
+      @review ? @review.update(r) : @review = Review.new(r)
+      if @review.save
+        render :show
+      else
+        render json: @review.errors.full_messages, status: 422
+      end
     else
-      render json: @review.errors.full_messages, status: 422
+      render json: ["Invalid user"], status: 403
     end
   end
 
@@ -33,10 +28,14 @@ class Api::ReviewsController < ApplicationController
   def destroy
     @review = Review.find_by(id: params[:id])
     if @review
-      if @review.destroy
-        render json: {}
+      if @review.user_id == current_user.id
+        if @review.destroy
+          render json: {}
+        else
+          render json: @review.errors.full_messages, status: 403
+        end
       else
-        render json: @review.errors.full_messages, status: 403
+        render json: ["Invalid user"], status: 403
       end
     else
       render json: ["Review not found"], status: 404
